mon_db.sh

+++ mon_db.sh
@@ -30,7 +30,7 @@ for HOST in $(xargs -n1 echo <<< "$HOSTS"); do
   echo "HOST="$HOST
   
 #  $BASEDIR/test_ssh.sh $CLIENT $HOST
-#  if [ "$?" -ne 0 ]; then echo "test_ssh.sh not return 0, continue"; continue; fi
+  if [ "$?" -ne 0 ]; then echo "test_ssh.sh not return 0, skip host: "$HOST", continue to next HOST"; continue; fi
   DBS=$($BASEDIR/iniget.sh $CONFIG $HOST db)
   for DB in $(xargs -n1 echo <<< "$DBS"); do
     echo "DB="$DB
@@ -51,17 +51,19 @@ for HOST in $(xargs -n1 echo <<< "$HOSTS"); do
 
     LOGF=$LOGDIR/mon_db_${HOST}_${DB}.log
     $WRTPI $HOST $DB db | sed -n '/V$INSTANCE:/{p;n;p;n;n;p;}' > $LOGF
+    
+    if [[ -s $LOGF ]]; then
+      LOGF_DB_DIFF=$LOGDIR/mon_db_${HOST}_${DB}_db_diff.log
+      LOGF_DB_OLD=$LOGDIR/mon_db_${HOST}_${DB}_db_old.log
+      touch $LOGF_DB_OLD
+      diff $LOGF_DB_OLD $LOGF > $LOGF_DB_DIFF
 
-    LOGF_DB_DIFF=$LOGDIR/mon_db_${HOST}_${DB}_db_diff.log
-    LOGF_DB_OLD=$LOGDIR/mon_db_${HOST}_${DB}_db_old.log
-    touch $LOGF_DB_OLD
-    diff $LOGF_DB_OLD $LOGF > $LOGF_DB_DIFF
-
-    if [ -s $LOGF_DB_DIFF ]; then 
-        cat $LOGF_DB_DIFF | $BASEDIR/send_msg.sh $CONFIG $SCRIPT_NAME $HOST $DB "DB status has been changed:"
+      if [ -s $LOGF_DB_DIFF ]; then 
+          cat $LOGF_DB_DIFF | $BASEDIR/send_msg.sh $CONFIG $SCRIPT_NAME $HOST $DB "DB status has been changed:"
+      fi
+      cp $LOGF $LOGF_DB_OLD
+      rm $LOGF_DB_DIFF $LOGF
     fi
-    cp $LOGF $LOGF_DB_OLD
-    rm $LOGF_DB_DIFF $LOGF
   done
 done
